utils/image_generation/openai/requirements.txt
openai==1.40.0
python-dotenv==1.0.1